EH-964: Moved DegreeInfo from Opiskelusuunnitelma to Tavoitteet
